requirements.txt
rdflib
requests
PyYAML
